+++ lib/src/fortnite_api_io/structures/challenges.dart
@@ -1,8 +1,15 @@
 /// fortnite challenges response
 class ChallengeResponse {
+  /// result success of response
   late bool result = false;
+
+  /// season number for quest bundles
   late int season = 0;
+
+  /// language of response
   late String lang = "en";
+
+  /// list of quest bundles
   List<ChallengeBundle> bundles = [];
 
   ChallengeResponse(dynamic res) {
@@ -33,14 +40,31 @@ class ChallengeResponse {
 
 /// fortnite challenge bundle
 class ChallengeBundle {
+  /// bundle id
   late String id;
+
+  /// bundle name
   late String name;
+
+  /// unlock type
   late String unlockType;
+
+  /// tags
   late List<dynamic> tags;
+
+  /// api tags
   late List<dynamic> apiTags;
+
+  /// colors for the bundle
   late dynamic colors;
+
+  /// images for quest bundle
   late dynamic images;
+
+  /// rewards for the bundle
   late List<dynamic> bundleRewards;
+
+  /// list of challenges
   List<Quest> quests = [];
 
   ChallengeBundle({
@@ -75,15 +99,34 @@ class ChallengeBundle {
 
 /// fortnite quest
 class Quest {
+  /// quest id
   late String id;
+
+  /// quest name
   late String name;
+
+  /// is the quest enabled
   late bool enabled;
+
+  /// date when quest was enabled
   late String enabledDate;
+
+  /// parest quest of the quest
   late dynamic parentQuest;
+
+  /// total progress of the quest
   late int progressTotal;
+
+  /// npc character of the quest
   late dynamic tandemCharacter;
+
+  /// tags of the quest
   late List<dynamic> tags;
+
+  /// reward object for quest
   late dynamic reward;
+
+  /// map locations for the quest
   late List<dynamic> locations;
 
   Quest({